refactor: subscription page nav tabs linked with react router
