+++ client/src/Router.tsx
@@ -5,7 +5,7 @@ import { UsersPage } from './pages/Users';
 import { TimetablePage } from './pages/Timetable';
 import { LessonsPage } from './pages/Lessons';
 import { StudentsPage } from './pages/Students';
-import { SubscriptionsPage } from './pages/Subscriptions';
+import { SubscriptionsPageLayout, SubscriptionsTemplatePage } from './pages/Subscriptions';
 
 function Hello() {
   return <h1>Work in progress...</h1>;
@@ -21,7 +21,12 @@ export function Router() {
         <Route path='/timetable' element={<TimetablePage />} />
         <Route path='/students' element={<StudentsPage />} />
         <Route path='/lessons' element={<LessonsPage />} />
-        <Route path='/subscriptions' element={<SubscriptionsPage />} />
+
+        <Route path='/subscriptions' element={<SubscriptionsPageLayout />}>
+          <Route index element={<Hello />} />
+          <Route path=':templates' element={<SubscriptionsTemplatePage />} />
+        </Route>
+
         <Route path='/finance' element={<Hello />} />
       </Route>
     </Routes>
