+++ client/src/pages/Subscriptions/Subscription/Page.tsx
@@ -1,0 +1,11 @@
+import { SubscriptionsHeader } from './PageHeader';
+import { SubscriptionContent } from './Content';
+
+export function SubscriptionsPage() {
+  return (
+    <>
+      <SubscriptionsHeader />
+      <SubscriptionContent />
+    </>
+  );
+}
