+++ client/src/pages/Subscriptions/Templates/Page.tsx
@@ -1,0 +1,11 @@
+import { SubscriptionsTemplatesHeader } from './PageHeader';
+import { SubscriptionsTemplatesContent } from './Content';
+
+export function SubscriptionsTemplatePage() {
+  return (
+    <>
+      <SubscriptionsTemplatesHeader />
+      <SubscriptionsTemplatesContent />
+    </>
+  );
+}
