+++ client/src/pages/Subscriptions/index.tsx
@@ -1,1 +1,3 @@
-export { SubscriptionsPage } from './Page';
+export { SubscriptionsPageLayout } from './PageLayout';
+export { SubscriptionsPage } from './Subscription/Page';
+export { SubscriptionsTemplatePage } from './Templates/Page';
